Use bash test script on windows CI tests as well Move comm base test class into test_CommBase to avoid problem with collecting tests from a module Catch pika.exceptions.StreamLostError when closing channel Don't do anything for RMQComm if purge is called after close Copy testing_options entry when getting test instance keyword arguments to prevent modification for later tests Add debug messages and display test output for MPI tests
