+++ .github/workflows/test-install.yml
@@ -258,10 +258,12 @@ jobs:
       if: ${{ contains(matrix.test-flags1, '--write-script') }}
       id: test1_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (1st)
       if: ${{ steps.test1_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Run tests (2nd)
       if: ${{ matrix.test-flags2 != '' }}
       id: test2
@@ -276,10 +278,12 @@ jobs:
         }}
       id: test2_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (2nd)
       if: ${{ steps.test2_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Combine coverage reports
       continue-on-error: true
       run: '# ls .*
@@ -563,10 +567,12 @@ jobs:
       if: ${{ contains(matrix.test-flags1, '--write-script') }}
       id: test1_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (1st)
       if: ${{ steps.test1_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Run tests (2nd)
       if: ${{ matrix.test-flags2 != '' }}
       id: test2
@@ -581,10 +587,12 @@ jobs:
         }}
       id: test2_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (2nd)
       if: ${{ steps.test2_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Combine coverage reports
       continue-on-error: true
       run: '# ls .*
@@ -844,10 +852,12 @@ jobs:
       if: ${{ contains(matrix.test-flags1, '--write-script') }}
       id: test1_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (1st)
       if: ${{ steps.test1_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Run tests (2nd)
       if: ${{ matrix.test-flags2 != '' }}
       id: test2
@@ -862,10 +872,12 @@ jobs:
         }}
       id: test2_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (2nd)
       if: ${{ steps.test2_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Combine coverage reports
       continue-on-error: true
       run: '# ls .*
@@ -1127,10 +1139,12 @@ jobs:
       if: ${{ contains(matrix.test-flags1, '--write-script') }}
       id: test1_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (1st)
       if: ${{ steps.test1_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Run tests (2nd)
       if: ${{ matrix.test-flags2 != '' }}
       id: test2
@@ -1145,10 +1159,12 @@ jobs:
         }}
       id: test2_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (2nd)
       if: ${{ steps.test2_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Combine coverage reports
       continue-on-error: true
       run: '# ls .*
@@ -1409,10 +1425,12 @@ jobs:
       if: ${{ contains(matrix.test-flags1, '--write-script') }}
       id: test1_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (1st)
       if: ${{ steps.test1_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Run tests (2nd)
       if: ${{ matrix.test-flags2 != '' }}
       id: test2
@@ -1427,10 +1445,12 @@ jobs:
         }}
       id: test2_script
       run: ./run_mpi.sh
+      shell: bash -l {0}
       continue-on-error: true
     - name: Re-run test script failures (2nd)
       if: ${{ steps.test2_script.outcome == 'failure' }}
       run: ./run_mpi.sh --last-failed --last-failed-no-failures=none --second-attempt
+      shell: bash -l {0}
     - name: Combine coverage reports
       continue-on-error: true
       run: '# ls .*
