+++ tests/conftest.py
@@ -172,6 +172,7 @@ def pytest_cmdline_preparse(args, dont_exit=False):
             if '--with-mpi' not in args:
                 args.append('--with-mpi')
             args += ['-p', 'no:flaky']
+            args += ['-s', '-o', 'log_cli=true']  # TODO: remove this
             for x in ['--reruns=2', '--reruns-delay=1', '--timeout=900']:
                 if x in args:
                     args.remove(x)
@@ -478,10 +479,10 @@ def write_pytest_script(fname, argv):
     cmd = ' '.join(argv)
     if platform._is_win:
         cmd = cmd.replace('\\', '/')
-        lines = [cmd + ' %*']
-    else:
-        lines = ['#!/bin/bash',
-                 cmd + ' $@']
+    #     lines = [cmd + ' %*']
+    # else:
+    lines = ['#!/bin/bash',
+             cmd + ' $@']
     with open(fname, 'w') as fd:
         fd.write('\n'.join(lines))
     os.chmod(fname, (stat.S_IRWXU
@@ -1406,14 +1407,19 @@ def sync_mpi_result(request, on_mpi):
     r"""Synchronize results between MPI ranks."""
     mpi_exchange = None
     if on_mpi:
+        print("new mpi exchange", request.node)
         mpi_exchange = new_mpi_exchange()
+        print("before sync in setup", mpi_exchange.global_tag, request.node)
         mpi_exchange.sync()
+        print("after sync in setup", mpi_exchange.global_tag, request.node)
     yield
     if on_mpi:
+        print("before sync in teardown", mpi_exchange.global_tag, request.node)
         failure = (request.node.rep_setup.failed
                    or getattr(getattr(request.node, 'rep_call', None),
                               'failed', False))
         mpi_exchange.finalize(failure)
+        print("after sync in teardown", mpi_exchange.global_tag, request.node)
 
 
 # Monkey patch pytest-cov plugin with MPI Barriers to prevent multiple
